Revert "Merge remote-tracking branch 'upstream/6.4' into cz-trans-review"
This reverts commit 4480452daccdfd69e53a94b50df84f15579bf6f2, reversing
changes made to a4dabd10694a7c81e627d4435981f35218f18d3e.

+++ src/Symfony/Component/Messenger/Bridge/Beanstalkd/Transport/Connection.php
@@ -123,7 +123,7 @@ class Connection
             $job = $this->client->useTube($this->tube)->put(
                 $message,
                 PheanstalkInterface::DEFAULT_PRIORITY,
-                (int) ($delay / 1000),
+                $delay / 1000,
                 $this->ttr
             );
         } catch (Exception $exception) {
